requirements.txt
langchain
langchain-community
langchain-mistralai
python-dotenv
streamlit
langchain-experimental
sentence-transformers
langchain-chroma
langchainhub
unstructured